Added more files and modified old ones

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rtammi <rtammi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 13:50:13 by rtammi            #+#    #+#             */
-/*   Updated: 2024/04/18 19:43:50 by rtammi           ###   ########.fr       */
+/*   Updated: 2024/04/18 20:39:58 by rtammi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
